+++ auto_database_backup/doc/RELEASE_NOTES.md
@@ -21,3 +21,27 @@
 #### UPDT
 
 - Fixed the errors while inputting list_db = False in odoo conf file.
+
+## Module <auto_database_backup>
+
+#### 21.05.2024
+#### Version 17.0.4.0.1
+#### UPDT
+
+- Changed response.list_buckets() function to head_bucket().
+
+## Module <auto_database_backup>
+
+#### 08.07.2024
+#### Version 17.0.5.0.1
+#### UPDT
+
+- Fixed the nextcloud auto remove backup error.
+
+## Module <auto_database_backup>
+
+#### 08.07.2024
+#### Version 17.0.6.0.1
+#### UPDT
+
+- Fixed the windows backup error.